ensure shakenfist servers have time service
Shakenfist servers, especially those running etcd, should have accurate
time. Clock skew can also be a problem with VMs.

This patch ensures that a time service is installed and running on all
servers. By default this is chrony which is modern and much better at
handling large initial skews than ntpd and thus does not also require
ntpdate. Using systemd-timesyncd is probably better, but it's not
available on Ubuntu Bionic LTS.

The package to install is included in the 'shakenfist_deps' variable and
the service as 'shakenfist_service_time' in case it needs to be
overriden (for example if you prefer to use systemd-timesyncd or even
ntpd).