support config of KSM (kernel same-page merging)
On some machines, especially when using nested virtualisation, KSM can
cause problems including soft CPU lockups.

This patch allows users to explicitly enable or disable KSM support on
shakenfist nodes. If defined, this will modify the /etc/default/qemu-kvm
config file and restart the qemu-kvm service if necessary.

A new variable "shakenfist_ksm" is introduced to allow this to be
configured. By default this will be set to "AUTO" which is the Ubuntu
default, but can be set to 0 (disabled) or 1 (enabled).

Users can also release existing pages and temporarily disable KSM with the
following commands:

 $ echo 2 |sudo tee /sys/kernel/mm/ksm/run
 $ echo 0 |sudo tee /sys/kernel/mm/ksm/run